[REFACTOR] ControllerAdvice를 추가합니다
1. UserController와 ArticleController에서 발생하는 예외에 대한 처리를 ControllerAdvice에서
처리해주도록 수정합니다.
2. 예외 상황인 경우 에러 페이지를 보여줄 수 있도록 error.html 파일을 추가하였습니다.
