+++ src/main/resources/templates/error.html
@@ -44,7 +44,7 @@
                                                 <div class="pdd-horizon-30 pdd-vertical-30">
                                                     <div class="mrg-btm-30">
                                                         <h4 class="text-center no-mrg-vertical mrg-btm-15 text-dark">
-                                                            페이지를 찾을 수 없습니다.</h4>
+                                                            Wrong access</h4>
                                                     </div>
                                                     <form action="/" method="get">
                                                         <div class="mrg-top-20 text-right">
